dcp-t5(001): move mcp/discoverygen under factory_sessions
Relocate pkg/transports/mcp/discoverygen to
pkg/services/factory_sessions/transports/mcp/discoverygen. The package
imports only factory_sessions MCP contracts, so it belongs in that
service's own transports subtree per the packaged-service-structure
convention.

Pure relocation: file contents are unchanged apart from the self-import
path in generate_test.go. Four external importers are retargeted. The
unit coverage minimum (82.68) moves with the package and the manifest is
re-sorted.
