+++ cmd/mcpdiscoverygen/main_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/portpowered/infinite-you/pkg/platform/generatedartifacts"
-	"github.com/portpowered/infinite-you/pkg/transports/mcp/discoverygen"
+	"github.com/portpowered/infinite-you/pkg/services/factory_sessions/transports/mcp/discoverygen"
 )
 
 type testFileSystem struct{}
