+++ pkg/services/factory_sessions/transports/mcp/discoverygen/generate.go
