+++ pkg/services/factory_sessions/transports/mcp/discoverygen/project.go
