+++ pkg/services/factory_sessions/transports/mcp/discoverygen/types.go
