+++ pkg/services/factory_sessions/transports/mcp/discoverygen/verify.go
